fix(core-manager): dev core manager
