+++ crates/core-manager/src/lib.rs
@@ -1,12 +1,27 @@
 #![feature(slice_take)]
 
 extern crate core;
-mod core_range;
+
 pub mod errors;
-pub mod manager;
+
 pub mod types;
+
+mod core_range;
+
+mod dev;
+
+mod dummy;
+
+mod manager;
+mod persistence;
+mod strict;
 
 pub use ccp_shared::types::CUID;
 pub use core_range::CoreRange;
 pub use cpu_utils::LogicalCoreId;
 pub use cpu_utils::PhysicalCoreId;
+pub use dev::DevCoreManager;
+pub use dummy::DummyCoreManager;
+pub use manager::CoreManager;
+pub use manager::CoreManagerFunctions;
+pub use strict::StrictCoreManager;
